requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
fastapi
uvicorn
sentence-transformers
chromadb
rank-bm25
pydantic
python-dotenv
openai
pypdf
python-multipart
python-docx